+++ crates/nu-engine/src/lib.rs
@@ -11,7 +11,7 @@ pub use column::get_columns;
 pub use documentation::get_full_help;
 pub use env::*;
 pub use eval::{
-    bind_args_to_stack, eval_block, eval_block_with_early_return, eval_call, eval_expression,
+    eval_block, eval_block_with_early_return, eval_call, eval_expression,
     eval_expression_with_input, eval_subexpression, eval_variable, redirect_env,
 };
 pub use glob_from::glob_from;
